HMAC verification working GET requests from  my Arduino (except have not tested timestamp parameter yet, as the Arduino can't do this natively - waiting for a real time clock part).
Basically it works, but there are some problems on the Arduino size due to resource constraints. Calculating the SHA256 HMAC does use a fair chunk of the Arduino's resources, and adding the SD card library to persistently record the counter on a micro SD card pushes it over the edge. Will look at a SHA1 option to see if its a more lightweight if theoretically less secure alternative (the problems with SHA1 or MD5 supposedly do not affect their use in HMACs).

An alternative to the counter and timestamp would be to record the HMAC values on the server, and use those to check for replay attacks. So long as the random string was unique this shouldn't cause a problem, however, the random number generator on the Arduino is awful. It reads values from a fixed set of random numbers, so if you start at the same place in the set the sequence is *always the same*.

git-svn-id: https://subversion.assembla.com/svn/impresscmsaddons/modules/straylight/trunk@24411 7a4d4d67-0c3a-0410-b24c-89d50c169223